+++ packages/tools/utils/src/[any-any]-[iterable-object].js
@@ -1,5 +1,5 @@
 /**
-* @module @svizzle/utils/function-[iterable-object]
+* @module @svizzle/utils/[any-any]-[iterable-object]
 */
 
 /**