+++ packages/tools/utils/src/string-[array-number].js
@@ -9,7 +9,7 @@ import * as _ from "lamb";
  * The same can be done by `arrayMaxWith(_.getKey(key))` but here we avoid invoking a function for all the items.
  * @see
  {@link module:@svizzle/utils/array-number.arrayMax|arrayMax},
- {@link module:@svizzle/utils/function-[array-number].arrayMaxWith|arrayMaxWith}
+ {@link module:@svizzle/utils/[any-number]:accumcb-[array-number].arrayMaxWith|arrayMaxWith}
  *
  * @function
  * @arg {string} key
@@ -39,7 +39,7 @@ export const arrayMaxBy = key => _.reduceWith((max, item) => {
  * The same can be done by `arrayMinWith(_.getKey(key))` but here we avoid invoking a function for all the items.
  * @see
  {@link module:@svizzle/utils/array-number.arrayMin|arrayMin},
- {@link module:@svizzle/utils/function-[array-number].arrayMinWith|arrayMinWith}
+ {@link module:@svizzle/utils/[any-number]:accumcb-[array-number].arrayMinWith|arrayMinWith}
  *
  * @function
  * @arg {string} key